Task managing, check permissions

+++ view/task/index.php
@@ -1,2 +1,2 @@
 <? $this->guard_user();
-$this->node('', array('user_id' => CMS\Me::id(), 'task' => CMS\Vars::URI('task')));
+$this->node('', array('user_id' => CMS\Me::id(), 'task' => (int) CMS\Vars::URI('task')));
